2_Embedding/requirements.txt
sentence_transformers
psycopg[binary]
pgvector
fastapi
uvicorn
tqdm
python-dotenv